third_party/WebKit/LayoutTests/inspector/sources/autocomplete-scss-expected.txt

The test verifies autocomplete suggestions for SCSS file.


Running: testPropertyValueSuggestionsBefore$
========= Selection In Editor =========
@mixin my-border-style($style) {
    border-style: |$;
}
======= Autocomplete Suggestions =======
[dashed, dotted, double, groove, hidden, inherit, initial, inset, none, outset, ridge, solid]

Running: testPropertyValueSuggestionsAfter$
========= Selection In Editor =========
@mixin my-border-style($style) {
    border-style: $|;
}
======= Autocomplete Suggestions =======
[$style]
